Improve mirroring code
Support slashes in index node values and document labels and also
support duplicate nodes values or documents labels. Slashes are
replaced with underscores. To handle duplicates, the primary key of
the object is appended to the label inside parenthesis. Closes
GitLab issue #722. Thanks to Rob de Canha-Knight (@rssfed23) for the
report and research.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>