Update workflow instance access requirements
Update the access requirements to match the intented layout and
the current REST API layout. In order to view the list of
log entries and the list of workflow instances of a document,
the workflow permission view is now required for the document
as well as the workflow template.

This double permission requirement now matches the same design
as the metadata and tags apps.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>