Drop the indexes from model renames
Django is unable to rename the named indexes when the
model is renamed. When another model is created or
renamed with the name of the original model, there
is an index name clash.

Directly drop the index prior to its recreation until
Django bug 23577 is fixed: https://code.djangoproject.com/ticket/23577

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>