Don't ignore chord task results
Change as indicated in the Celery documentation:
https://docs.celeryproject.org/en/latest/userguide/canvas.html#important-notes

Important Notes

Tasks used within a chord must not ignore their results. In practice this
means that you must enable a result_backend in order to use chords.
Additionally, if task_ignore_result is set to True in your configuration,
be sure that the individual tasks to be used within the chord are defined
with ignore_result=False. This applies to both Task subclasses and decorated tasks.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>