Add the organizations app
This app handle sdetails that are specific to an organizations installation
of Mayan EDMS. The app has two settings: ``ORGANIZATIONS_INSTALLATION_URL``
which controls the URL of the installation. This URL is used when composing
the document link URL when sending document links via email. It is also used
to create the download link for exported document versions.

If this setting is not specified, Mayan EDMS will try to obtain the URL
scheme, host, and port from the HTTP request itself. When running Mayan EDMS
behind proxy servers or load balancers this might not return accurate results
and will require an explicit value for the setting.

The setting ``ORGANIZATIONS_URL_BASE_PATH`` replaces the previous setting
named ``COMMON_URL_BASE_PATH`` used when hosting Mayan EDMS in a path other
than the root path ('/') of the web server.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>