Add text helper to convert lists to queryset
Used by tests that generate more than one test object.
If the instances `repr` results is the same, resulting
queryset will be sorted by ID and processed in a different
order than they were created and supplied to the view.

Converting the list to a queryset ensures the assertion stage
of the test has the same test object queryset as the view
it is testing.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>