Backport image generation retry setting
The setting is named `CONVERTER_IMAGE_GENERATION_MAX_RETRIES`
and it is the new setting to control the behavior globally in
version 4.1.

Instead of adding a hardcoded literal, this setting was backported
for easier transition into version 4.1.

This setting allows changing the image generation task maximum
retry count. Celery's built in default value is 3, this setting
increases that default to 7.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>