Fix issue creating OCR update workflow action
Fix issue when attempting to create a Document version page OCR
update workflow action. Instead of the model class, the template
form field now passes the ``app_label`` and the ``model_name`` of
the model via the widget attributes to avoid Django's attribute
template to attempt getting a string representation of the model.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>