Add permission filtering to source switch links
These are the dynamic navigation links that are shown
when creating a new document or uploading a new file
to an existing document.

The permission filtering will be the same as the views:
document create permission for the
source links during document uploads and document file new
permissions for the source links in the new document file
upload view.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>