Don't tag minor version as `latest`
Don't tag Docker images as `latest` for minor releases.

As per Docker's specifications, the `latest` tag is applied to
the latest image built if no tag is specified. It is not meant
to represent the latest version of an project, just the last
image that has been built.

However users commonly (and mistakenly) have come to expect the
`latest` tag to represent the latest version of the project.
The GitLab CI file is updated to fulfill this expectation.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>