Add a patch for Python's CVE-2007-4559
https://nvd.nist.gov/vuln/detail/CVE-2007-4559.

This is not a vulnerability of Mayan EDMS but of a Python standard
library that manages Tar archives. A malicious Tar files can contain
files whose paths are prefixed with double dot causing the act of
extracting the archive to place those files in directories other
than where the archive was meant to be extracted.

This patch adds a verification to ensure that the final computed
path of a file to be extracted is still relative to the extraction
path.

Exploiting this vulnerability for older versions of Mayan EDMS
requires compromising an existing package hosted on the NPM
registry and adding Python code specifically targeting Mayan EDMS.

As part of Mayan EDMS design philosophies, dependencies are only
downloaded from authoritative locations and more importantly, each
dependency is pinned to a specific version to guarantee immutable
releases.

Due to all these factors, the surface of attack of this
vulnerability is very limited for older versions of Mayan EDMS,
it is also very improbable, very difficulty to accomplish and
very difficult to remain undetected.

There are no known actual or theoretical attacks for Mayan EDMS
exploiting this vulnerability.

Thanks to the TrellixVulnTeam for the pull request which lead
to this Mayan EDMS specific patch.

Signed-off-by: Roberto Rosario <roberto.rosario@mayan-edms.com>